+++ packages/lib/src/mongoSchema/index.ts
@@ -1,3 +1,4 @@
 export type InterfaceExtended<T> = T & { _id: string; _doc?: T };
 
 export * from "./cle/classe";
+export * from "./cle/etablissement";
